build: Replace package.json and package-lock.json.
During the migration to Next.js, the package.json and package-lock.json generated by "npx create-next-app" were replaced by the same files generated from "npx create-react-app". Because of this, it was necessary to replace the current package.json and package-lock.json with those that would be suitable for the current state of the project.